added player movement and tested movement script
Player movement has been added. Currently the player only moves left or right using the arrow keys. It is far from ideal in terms of code. I will need to refactor the passing of player_direction to draw_player and improve that. I also need to refactor the player variables into a player object instead so that the code is nice and readable.